refactor(cli): move the handlers and their tests into directories
Twelve handlers, their Json records and their tests were all in the
project root, so the file list was two dozen entries deep before
reaching anything structural. Handlers/ and Tests/Handlers/ mirror each
other.
