+++ CodeWalker.Cli/ExportService.cs
@@ -5,6 +5,7 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
+using CodeWalker.Cli.Handlers;
 using CodeWalker.Cli.Helpers;
 using CodeWalker.GameFiles;
