+++ CodeWalker.Cli/Handlers/PackHandler.cs
@@ -8,7 +8,7 @@ using System.Threading;
 using CodeWalker.Cli.Helpers;
 using CodeWalker.GameFiles;
 
-namespace CodeWalker.Cli;
+namespace CodeWalker.Cli.Handlers;
 
 internal sealed record PackOptions
 {
