+++ CodeWalker.Cli/Tests/Handlers/DiffHandlerTests.cs
@@ -1,12 +1,12 @@
 using System;
 using System.IO;
 
+using CodeWalker.Cli.Handlers;
 using CodeWalker.Cli.Helpers;
 
 using Xunit;
-using Xunit.v3;
 
-namespace CodeWalker.Cli.Tests;
+namespace CodeWalker.Cli.Tests.Handlers;
 
 public sealed class DiffHandlerTests
 {
