+++ CodeWalker.Cli/Tests/Handlers/ExtractHandlerTests.cs
@@ -1,28 +1,35 @@
 using System;
 using System.IO;
 
+using CodeWalker.Cli.Handlers;
 using CodeWalker.Cli.Helpers;
 
 using Xunit;
-using Xunit.v3;
 
-namespace CodeWalker.Cli.Tests;
+namespace CodeWalker.Cli.Tests.Handlers;
 
 [Collection("ConsoleOutput")]
-public sealed class StatHandlerTests
+public sealed class ExtractHandlerTests
 {
-    private static RpfOptions MakeOptions(string rpfPath, bool json) =>
+    private static ExtractOptions MakeOptions(string rpfPath, bool json, bool dryRun = false) =>
         new()
         {
-            RpfPath = rpfPath,
-            ExePath = "/nonexistent",
-            Gen9 = false,
-            Filters = [],
-            Verbose = false,
-            Json = json,
-            Recursive = false,
-            Threads = 1,
-            SizeFormat = SizeFormat.IEC,
+            Rpf = new RpfOptions
+            {
+                RpfPath = rpfPath,
+                ExePath = "/nonexistent",
+                Gen9 = false,
+                Filters = [],
+                Verbose = false,
+                Json = json,
+                Recursive = false,
+                Threads = 1,
+                SizeFormat = SizeFormat.IEC,
+            },
+            OutputPath = "/tmp/cw_extract_out",
+            DryRun = dryRun,
+            NoOverwrite = false,
+            Progress = false,
         };
 
     // ── Validation failures ────────────────────────────────────────────
@@ -38,7 +45,7 @@ public sealed class StatHandlerTests
             Console.SetOut(new StringWriter());
             Console.SetError(stderr);
 
-            int exitCode = StatHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: false), TestContext.Current.CancellationToken);
+            int exitCode = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: false), TestContext.Current.CancellationToken);
 
             Assert.Equal(1, exitCode);
             Assert.Contains("Error:", stderr.ToString());
@@ -61,7 +68,7 @@ public sealed class StatHandlerTests
             Console.SetOut(stdout);
             Console.SetError(new StringWriter());
 
-            int exitCode = StatHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
+            int exitCode = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
 
             Assert.Equal(1, exitCode);
             string output = stdout.ToString();
@@ -84,15 +91,15 @@ public sealed class StatHandlerTests
             StringWriter stdout = new();
             Console.SetOut(stdout);
 
-            _ = StatHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
+            _ = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
 
             string output = stdout.ToString();
             Assert.Contains("\"rpfFile\":", output);
             Assert.Contains("\"totalFiles\": 0", output);
-            Assert.Contains("\"totalSize\": 0", output);
-            Assert.Contains("\"resourceCount\": 0", output);
-            Assert.Contains("\"binaryCount\": 0", output);
-            Assert.Contains("\"compressionRatio\": 0", output);
+            Assert.Contains("\"extracted\": 0", output);
+            Assert.Contains("\"skipped\": 0", output);
+            Assert.Contains("\"errors\": 0", output);
+            Assert.Contains("\"dryRun\": false", output);
         }
         finally { Console.SetOut(origOut); }
     }
@@ -100,7 +107,7 @@ public sealed class StatHandlerTests
     [Fact]
     public void Execute_MissingExe_WithExistingRpf_ReturnsOne()
     {
-        string dir = Path.Combine(Path.GetTempPath(), "cw_stat_" + Guid.NewGuid().ToString("N"));
+        string dir = Path.Combine(Path.GetTempPath(), "cw_ext_" + Guid.NewGuid().ToString("N"));
         _ = Directory.CreateDirectory(dir);
         string rpf = Path.Combine(dir, "test.rpf");
         File.WriteAllBytes(rpf, []);
@@ -114,7 +121,7 @@ public sealed class StatHandlerTests
                 StringWriter stderr = new();
                 Console.SetError(stderr);
 
-                int exitCode = StatHandler.Execute(MakeOptions(rpf, json: false), TestContext.Current.CancellationToken);
+                int exitCode = ExtractHandler.Execute(MakeOptions(rpf, json: false), TestContext.Current.CancellationToken);
 
                 Assert.Equal(1, exitCode);
                 Assert.Contains("Error:", stderr.ToString());
@@ -126,5 +133,22 @@ public sealed class StatHandlerTests
             }
         }
         finally { Directory.Delete(dir, true); }
+    }
+
+    [Fact]
+    public void Execute_DryRun_Json_ErrorStillHasDryRunTrue()
+    {
+        TextWriter origOut = Console.Out;
+        try
+        {
+            StringWriter stdout = new();
+            Console.SetOut(stdout);
+
+            _ = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true, dryRun: true), TestContext.Current.CancellationToken);
+
+            string output = stdout.ToString();
+            Assert.Contains("\"dryRun\": true", output);
+        }
+        finally { Console.SetOut(origOut); }
     }
 }
