+++ CodeWalker.Cli/Tests/Handlers/ListHandlerTests.cs
@@ -1,35 +1,28 @@
 using System;
 using System.IO;
 
+using CodeWalker.Cli.Handlers;
 using CodeWalker.Cli.Helpers;
 
 using Xunit;
-using Xunit.v3;
 
-namespace CodeWalker.Cli.Tests;
+namespace CodeWalker.Cli.Tests.Handlers;
 
 [Collection("ConsoleOutput")]
-public sealed class ExtractHandlerTests
+public sealed class ListHandlerTests
 {
-    private static ExtractOptions MakeOptions(string rpfPath, bool json, bool dryRun = false) =>
+    private static RpfOptions MakeOptions(string rpfPath, bool json) =>
         new()
         {
-            Rpf = new RpfOptions
-            {
-                RpfPath = rpfPath,
-                ExePath = "/nonexistent",
-                Gen9 = false,
-                Filters = [],
-                Verbose = false,
-                Json = json,
-                Recursive = false,
-                Threads = 1,
-                SizeFormat = SizeFormat.IEC,
-            },
-            OutputPath = "/tmp/cw_extract_out",
-            DryRun = dryRun,
-            NoOverwrite = false,
-            Progress = false,
+            RpfPath = rpfPath,
+            ExePath = "/nonexistent",
+            Gen9 = false,
+            Filters = [],
+            Verbose = false,
+            Json = json,
+            Recursive = false,
+            Threads = 1,
+            SizeFormat = SizeFormat.IEC,
         };
 
     // ── Validation failures ────────────────────────────────────────────
@@ -45,7 +38,7 @@ public sealed class ExtractHandlerTests
             Console.SetOut(new StringWriter());
             Console.SetError(stderr);
 
-            int exitCode = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: false), TestContext.Current.CancellationToken);
+            int exitCode = ListHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: false), TestContext.Current.CancellationToken);
 
             Assert.Equal(1, exitCode);
             Assert.Contains("Error:", stderr.ToString());
@@ -68,7 +61,7 @@ public sealed class ExtractHandlerTests
             Console.SetOut(stdout);
             Console.SetError(new StringWriter());
 
-            int exitCode = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
+            int exitCode = ListHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
 
             Assert.Equal(1, exitCode);
             string output = stdout.ToString();
@@ -91,15 +84,14 @@ public sealed class ExtractHandlerTests
             StringWriter stdout = new();
             Console.SetOut(stdout);
 
-            _ = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
+            _ = ListHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true), TestContext.Current.CancellationToken);
 
             string output = stdout.ToString();
             Assert.Contains("\"rpfFile\":", output);
             Assert.Contains("\"totalFiles\": 0", output);
-            Assert.Contains("\"extracted\": 0", output);
-            Assert.Contains("\"skipped\": 0", output);
-            Assert.Contains("\"errors\": 0", output);
-            Assert.Contains("\"dryRun\": false", output);
+            Assert.Contains("\"totalSize\": 0", output);
+            Assert.Contains("\"totalSizeFormatted\": \"0 B\"", output);
+            Assert.Contains("\"nestedRpfCount\": 0", output);
         }
         finally { Console.SetOut(origOut); }
     }
@@ -107,7 +99,7 @@ public sealed class ExtractHandlerTests
     [Fact]
     public void Execute_MissingExe_WithExistingRpf_ReturnsOne()
     {
-        string dir = Path.Combine(Path.GetTempPath(), "cw_ext_" + Guid.NewGuid().ToString("N"));
+        string dir = Path.Combine(Path.GetTempPath(), "cw_list_" + Guid.NewGuid().ToString("N"));
         _ = Directory.CreateDirectory(dir);
         string rpf = Path.Combine(dir, "test.rpf");
         File.WriteAllBytes(rpf, []);
@@ -121,7 +113,7 @@ public sealed class ExtractHandlerTests
                 StringWriter stderr = new();
                 Console.SetError(stderr);
 
-                int exitCode = ExtractHandler.Execute(MakeOptions(rpf, json: false), TestContext.Current.CancellationToken);
+                int exitCode = ListHandler.Execute(MakeOptions(rpf, json: false), TestContext.Current.CancellationToken);
 
                 Assert.Equal(1, exitCode);
                 Assert.Contains("Error:", stderr.ToString());
@@ -133,22 +125,5 @@ public sealed class ExtractHandlerTests
             }
         }
         finally { Directory.Delete(dir, true); }
-    }
-
-    [Fact]
-    public void Execute_DryRun_Json_ErrorStillHasDryRunTrue()
-    {
-        TextWriter origOut = Console.Out;
-        try
-        {
-            StringWriter stdout = new();
-            Console.SetOut(stdout);
-
-            _ = ExtractHandler.Execute(MakeOptions("/nonexistent/test.rpf", json: true, dryRun: true), TestContext.Current.CancellationToken);
-
-            string output = stdout.ToString();
-            Assert.Contains("\"dryRun\": true", output);
-        }
-        finally { Console.SetOut(origOut); }
     }
 }
